requirements.txt
python-telegram-bot==20.7
python-dateutil==2.8.2
python-dotenv==1.0.0